Implement a simple KV store
