+++ diypaxos/proto/simplekvstore.proto
@@ -1,0 +1,52 @@
+syntax = "proto3";
+
+import "google/rpc/status.proto";
+
+service SimpleKvStore {
+  rpc Get (GetRequest) returns (GetResponse) {}
+
+  rpc Insert (InsertRequest) returns (InsertResponse) {}
+
+  rpc Remove (RemoveRequest) returns (RemoveResponse) {}
+
+  rpc Update (UpdateRequest) returns (UpdateResponse) {}
+
+  rpc Upsert (UpsertRequest) returns (UpsertResponse) {}
+}
+
+message InsertRequest {
+  string key = 1;
+  bytes value = 2;
+}
+
+message InsertResponse {
+  google.rpc.Status status = 1;
+}
+
+message GetRequest {
+  string key = 1;
+}
+
+message GetResponse {
+  
+}
+
+message UpsertRequest {
+  string key = 1;
+  bytes value = 2;
+}
+
+message UpsertResponse {}
+
+message UpdateRequest {
+  string key = 1;
+  bytes value = 2;
+}
+
+message UpdateResponse {}
+
+message RemoveRequest {
+  string key = 1;
+}
+
+message RemoveResponse {}
