+++ diypaxos/server/BUILD
@@ -1,19 +1,13 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "server_lib",
-    srcs = ["main.go"],
+    name = "server",
+    srcs = ["server.go"],
     importpath = "diy-paxos/diypaxos/server",
-    visibility = ["//visibility:private"],
+    visibility = ["//visibility:public"],
     deps = [
         "//diypaxos/proto",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//reflection",
+        "@org_golang_google_grpc//codes",
+        "@org_golang_google_grpc//status",
     ],
-)
-
-go_binary(
-    name = "server",
-    embed = [":server_lib"],
-    visibility = ["//visibility:public"],
 )
