+++ go.mod
@@ -1,6 +1,6 @@
-module github.com/diy-paxos
+module diy-paxos
 
-go 1.18
+go 1.19
 
 require google.golang.org/grpc v1.51.0
